♻️ refactor: change count fields from Integer to Long in response

+++ src/main/java/com/honlife/core/app/controller/dashboard/payload/CategoryRankResponse.java
@@ -1,15 +1,16 @@
 package com.honlife.core.app.controller.dashboard.payload;
 
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
 public class CategoryRankResponse {
 
     // 순위
-    private Integer rank;
+    private Long rank;
     // 카테고리 이름
     private String categoryName;
     // 완료한 카테고리 수
-    private Integer completedCount;
+    private Long completedCount;
 
 }
